requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.6.0
arabic-reshaper==3.0.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2023.7.22
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.10
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.3
django-allauth==65.3.0
django-crispy-forms==2.3
django-hashid-field==3.4.1
django-redis==5.4.0
django-widget-tweaks==1.5.0
django-zarinpal==0.1.5
djangorestframework==3.15.2
frozenlist==1.5.0
hashids==1.3.1
idna==3.4
isodate==0.7.2
joblib==1.4.2
lxml==5.3.0
maxminddb==2.6.2
multidict==6.1.0
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.1
pyaes==1.6.1
pyasn1==0.6.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.10.1
pyTelegramBotAPI==4.24.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
redis==5.2.1
regex==2024.11.6
reportlab==4.3.0
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
setuptools==75.6.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.2
subprocess.run==0.0.8
Telethon==1.38.1
typing_extensions==4.8.0
tzdata==2024.2
urllib3==2.1.0
wheel==0.45.1
yarl==1.18.3
zarinpal==1.0.0
zeep==4.3.1
uwsgi
